+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix README for combined custom backend bootstrap
 - Allow annotation `@odata.type` in lower case format
 - Allow type prefix `datetime` in addition to `datetimeoffset`
+- Add peer dependency @types/express
 
 ## Version 1.8.5 - 2022-01-17