+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- New version information is automatically added after line 8 -->
 
-## Version 1.8.8 - tbd
+## Version 1.8.8 - 2022-02-08
 
 ### Fixed
+
+- Proxy option `calcContentDisposition` to calculate `content-disposition` header even if already available
 
 ## Version 1.8.7 - 2022-02-03